Move channels.ts to agile folder
The channels.ts file is created with the agile workflow in mind.
Therefore it belongs to the agile subdirectory and it is not a general utility file.
