+++ api/src/discord/agile/commands/workingOn.ts
@@ -9,7 +9,7 @@ import {
   sendStopWorkingOnMessage,
 } from "../../../plugins/discordHooks";
 import { getUserByDiscordId } from "../../database/users";
-import { getCurrentTaskChannelFromDiscord } from "../../utils/channels";
+import { getCurrentTaskChannelFromDiscord } from "../channels";
 
 async function accessDenied(interaction: CommandInteraction) {
   await interaction.editReply({
